+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/SignalingChannel.kt
@@ -27,6 +27,7 @@ import java.net.InetSocketAddress
 import java.net.Proxy
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicBoolean
+import java.util.concurrent.atomic.AtomicReference
 
 interface SignalingChannel {
 
@@ -40,7 +41,7 @@ interface SignalingChannel {
 
     interface Listener {
         fun onConnect(endpoint: String)
-        fun onDisconnect(disconnectReason: SoraDisconnectReason?)
+        fun onDisconnect(disconnectReason: SoraDisconnectReason?, result: SignalingChannelDisconnectResult?)
         fun onInitialOffer(offerMessage: OfferMessage, endpoint: String)
         fun onSwitched(switchedMessage: SwitchedMessage)
         fun onUpdatedOffer(sdp: String)
@@ -143,6 +144,8 @@ class SignalingChannelImpl @JvmOverloads constructor(
 
     private val receivedRedirectMessage = AtomicBoolean(false)
 
+    private val signalingChannelDisconnectResult = AtomicReference<SignalingChannelDisconnectResult?>()
+
     override fun connect() {
         SoraLogger.i(TAG, "[signaling:$role] endpoints=$endpoints")
         synchronized(this) {
@@ -240,7 +243,7 @@ class SignalingChannelImpl @JvmOverloads constructor(
 
         // type: redirect を受信している場合は onDisconnect を発火させない
         if (!receivedRedirectMessage.get()) {
-            listener?.onDisconnect(disconnectReason)
+            listener?.onDisconnect(disconnectReason, signalingChannelDisconnectResult.get())
         }
         listener = null
     }
@@ -476,6 +479,8 @@ class SignalingChannelImpl @JvmOverloads constructor(
                 SoraLogger.d(TAG, "[signaling:$role] @onClosed: skipped")
                 return
             }
+
+            signalingChannelDisconnectResult.set(SignalingChannelDisconnectResult(code, reason))
 
             try {
                 if (code != 1000) {