+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/SignalingChannelDisconnectResult.kt
@@ -1,0 +1,6 @@
+package jp.shiguredo.sora.sdk.channel.signaling
+
+data class SignalingChannelDisconnectResult(
+    val code: Int,
+    val reason: String,
+)